+++ src/MainPage/catalog-genres-list.tsx
