+++ src/MainPage/film-card.tsx
