+++ src/MoviePage/movie-page.tsx
@@ -1,5 +1,5 @@
-import {Footer} from '../Footer';
-import {Header} from '../Header';
+import {Footer} from '../footer';
+import {Header} from '../header';
 
 export const MoviePage = () => (
   <div>
