+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import {App} from './App';
+import {App} from './app';
 import {catalogFilmCards, filmCardData} from './data/data';
 
 const root = ReactDOM.createRoot(
